Move vecbuf tests to their own module to make panic triage easier.
The traige process skips files named "*_test.rs".
